move logic to conversationPartner component
